restart+score added
