+++ game2048/components/Header/Header.tsx
@@ -2,14 +2,20 @@ import styles from '../Header/Header.module.scss';
 import Restart from '../Restart/Restart';
 import Score from '../Score/Score';
 
-const Header = () => {
+interface HeaderProps {
+  score: number;
+  numbers: number[][];
+  restart: () => void;
+}
+
+const Header = (props: HeaderProps) => {
   return (
     <div className={styles.wrapper}>
       <div className={styles.container}>
         <div className={styles.title}>2048</div>
         <div className={styles.menu}>
-          <Restart />
-          <Score />
+          <Restart restart={() => props.restart()} numbers={props.numbers} />
+          <Score score={props.score} />
         </div>
       </div>
       <div className={styles.description}>Join the numbers and get to the 2048 tile!</div>
