+++ game2048/components/Score/Score.tsx
@@ -1,10 +1,14 @@
 import styles from '../Score/Score.module.scss';
 
-const Score = () => {
+interface ScoreProps {
+  score: number;
+}
+
+const Score = (props: ScoreProps) => {
   return (
     <div className={styles.container}>
       <div className={styles.title}>SCORE</div>
-      <div className={styles.number}>12345</div>
+      <div className={styles.number}>{props.score}</div>
     </div>
   );
 };
